+++ morebuiltins/cmd/log_server.py
@@ -1,7 +1,6 @@
 import asyncio
 import json
 import logging
-import logging.handlers
 import os
 import signal
 import sys